Move MovieQueryTask into a new file.
